Remove `Authenticators` controller concern
